requirements.txt
asgiref==3.3.1
attrs==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.28
django-cors-headers==3.1.1
djangorestframework==3.12.4
drf-yasg==1.17.1
idna==2.5
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jsonschema==4.0.0
MarkupSafe==1.1.1
packaging==20.9
pycodestyle==2.8.0
pyparsing==3.0.0rc1
pyrsistent==0.17.3
pytz==2022.6
PyYAML==5.3.1
requests==2.25.1
ruamel.yaml==0.17.21
sqlparse==0.4.3
typing_extensions==3.10.0.2
tzdata==2022.6
uritemplate==3.0.1
urllib3==1.26.2
